requirements.txt
customtkinter==4.5.10
darkdetect==0.7.1
et-xmlfile==1.1.0
future==0.18.2
iso8601==1.0.2
numpy==1.23.2
openpyxl==3.0.10
pandas==1.4.4
Pillow==9.2.0;
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
